[ chore ] 리뉴얼 페이지로 메인페이지 index설정, 기존 메인화면 코드 삭제
