+++ apps/desktop/src/@components/about/index.tsx
@@ -1,21 +1,21 @@
-import styled from "styled-components";
-import Footer from "../@common/footer";
-import { useRef } from "react";
-import AboutUs from "./aboutUs";
-import HowToUse from "./howToUse";
-import AboutMain from "./aboutMain";
-import MainHeader from "../main/renewal/mainHeader";
+import styled from 'styled-components';
+import Footer from '../@common/footer';
+import { useRef } from 'react';
+import AboutUs from './aboutUs';
+import HowToUse from './howToUse';
+import AboutMain from './aboutMain';
+import MainHeader from '../main/mainHeader';
 
 export default function About() {
   const aboutSectionRef = useRef<HTMLTableSectionElement>(null);
   const howToSectionRef = useRef<HTMLTableSectionElement>(null);
 
   function handleMoveAboutSection() {
-    aboutSectionRef.current?.scrollIntoView({ behavior: "smooth" });
+    aboutSectionRef.current?.scrollIntoView({ behavior: 'smooth' });
   }
 
   function handleMoveHowToSection() {
-    howToSectionRef.current?.scrollIntoView({ behavior: "smooth" });
+    howToSectionRef.current?.scrollIntoView({ behavior: 'smooth' });
   }
 
   return (
