+++ apps/desktop/src/@components/event/index.tsx
@@ -1,12 +1,12 @@
-import { useLocation } from "react-router-dom";
-import Footer from "../@common/footer";
-import MainHeader from "../main/renewal/mainHeader";
-import EventList from "./eventList";
-import HotEvent from "./hotEvent";
-import EventDetail from "./eventDetail";
-import styled from "styled-components";
-import { ScrollTopBtnIc } from "../../assets";
-import { useRef } from "react";
+import { useLocation } from 'react-router-dom';
+import Footer from '../@common/footer';
+import MainHeader from '../main/mainHeader';
+import EventList from './eventList';
+import HotEvent from './hotEvent';
+import EventDetail from './eventDetail';
+import styled from 'styled-components';
+import { ScrollTopBtnIc } from '../../assets';
+import { useRef } from 'react';
 
 export default function Event() {
   const pathname = useLocation().pathname;
@@ -14,13 +14,13 @@ export default function Event() {
   const hotEventSectionRef = useRef<HTMLTableSectionElement>(null);
 
   function handleMoveHotEventSection() {
-    hotEventSectionRef.current?.scrollIntoView({ behavior: "smooth" });
+    hotEventSectionRef.current?.scrollIntoView({ behavior: 'smooth' });
   }
 
   return (
     <>
       <MainHeader />
-      {pathname === "/event" ? (
+      {pathname === '/event' ? (
         <>
           <HotEvent scrollRef={hotEventSectionRef} />
           <EventList />
