+++ apps/desktop/src/@components/main/recentTrackList.tsx
@@ -1,14 +1,14 @@
-import styled from "styled-components";
-import { useGetRecentTracks } from "../../../hooks/queries/tracks";
-import { FilteredTrackType } from "../../../type/tracks";
-import { useState } from "react";
-import RecentTrackItem from "./recentTrackItem";
+import styled from 'styled-components';
+import { useGetRecentTracks } from '../../hooks/queries/tracks';
+import { FilteredTrackType } from '../../type/tracks';
+import { useState } from 'react';
+import RecentTrackItem from './recentTrackItem';
 
 export default function RecentTrackList() {
   const { recentTrackInfo } = useGetRecentTracks(4);
-  const [playingTrack, setPLayingTrack] = useState<FilteredTrackType["trackId"] | null>(null);
+  const [playingTrack, setPLayingTrack] = useState<FilteredTrackType['trackId'] | null>(null);
 
-  function selectTrack(trackId: FilteredTrackType["trackId"]) {
+  function selectTrack(trackId: FilteredTrackType['trackId']) {
     setPLayingTrack(trackId);
   }
