--- apps/desktop/src/@pages/homePage.tsx
@@ -1,5 +1,0 @@
-import Home from "../@components/main/renewal/home";
-
-export default function HomePage() {
-  return <Home />;
-}
